refactor(BSP-37): removed hook refactor
