+++ src/pages/PlacesMap.tsx
@@ -1,17 +1,7 @@
 import DisplayMap from '../components/map/DisplayMap';
-import { useSpeciesMapUrl } from '../components/map/useSpeciesMapUrl';
 
 const PlacesMap = () => {
-  const { speciesMapUrl, isFetching, invalidateMapToken } = useSpeciesMapUrl();
-
-  return (
-    <DisplayMap
-      height={'100%'}
-      src={speciesMapUrl}
-      isFetching={isFetching}
-      onInvalidToken={invalidateMapToken}
-    />
-  );
+  return <DisplayMap height={'100%'} />;
 };
 
 export default PlacesMap;
